frontend/requirements.txt
streamlit==1.29.0
streamlit-option-menu==0.3.6
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.1
numpy==1.26.0
plotly==5.17.0
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
pyjwt==2.8.0
pydantic==2.4.2
python-multipart==0.0.6
email-validator==2.1.0.post131.0
streamlit-aggrid==0.3.4
streamlit-echarts==0.4.0
folium==0.15.1
streamlit-folium==0.17.0
pydantic==2.0.0
